Replaced conformation popup with terms and cond. (#101)
* Replaced conformation popup with terms and cond.
Replaced the conformation popup after making a cart that prompted for confirmation or edits with the Statement Of Understanding
Functionally the same, but uses comment cupoard's Statement of Understanding component instead of a custom popup.
I also added a space in the inventory management page between the review change and to top button to make it look better.

* Should fix the issue by moving when API is called
I am uncertain why my previous commit worked on my instance and not on other instances
I am more confident of this comit but given the previous comit worked as well, I am slightly uncertain
As long as it is run, then it works well

I have only slightly changed how the button in the CartView sidebar is displayed. If a custom button text is desired, I can do that, I want to be sure that my current comit works first

* fix retracting cart

---------

Co-authored-by: Dennis Keithly <dxk200017@utdallas.edu>
Co-authored-by: Daniel Nguyen <94246004+CrazEpic@users.noreply.github.com>