Fixed the styling issues in the Admin Dashboard and Queue pages.
Reworked the submit button on the Edit Source page to the back arrow button and the delete button on the same page to the trashcan button next to the source name.

Reworked pages to use path parameters instead of history.

Reworked the cancel button on the delete source page to redirect back to the Edit Source page instead of the main Sources page.